// webui/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Inspection: {
            format?: string | null;
            version?: string | null;
            platforms?: components["schemas"]["Platform"][];
            cores?: components["schemas"]["InspectedCore"][];
            plugins?: components["schemas"]["InspectedPlugin"][];
            configs?: components["schemas"]["ShippedConfigFile"][];
            release?: components["schemas"]["BundleRelease"];
            warnings?: components["schemas"]["Finding"][];
        };
// ...
        ShippedConfigFile: {
            file?: string;
            /** @description The plugin it belongs to */
            plugin?: string | null;
            params?: components["schemas"]["ConfigParam"][];
        };
// ...
    };
// ...
}